requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-ckeditor==6.1.0
django-colorfield==0.4.2
django-cors-headers==3.8.0
django-cropperjs==0.0.5
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-modeltranslation==0.17.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
iyzipay==1.0.19
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
